birds/birds/requirements.txt
django==3.1.8
django-extensions
djangorestframework
sqlparse
markdown